+++ resources/js/Components/sales/SaleHistory.js
@@ -97,7 +97,7 @@ const SaleHistory = ({ sale, saleDetails }) => {
                                 </td>
                                 <td className="py-4 px-6 text-sm font-medium text-gray-900 whitespace-nowrap ">
                                     {formatCurrency(
-                                        parseInt(detail.sales_price) *
+                                        parseFloat(detail.sales_price) *
                                             detail.amount
                                     )}
                                 </td>
